# docs/openapi-phase-6.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Person Search Phase 6 API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "session": []
    }
  ],
  "paths": {
    "/cases": {
      "get": {
        "summary": "List Cases (Operator: own only, Viewer: all)",
        "parameters": [
          {
            "name": "owner_user_id",
            "in": "query",
            "description": "Viewer only; ignored for Operator",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "created_from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "created_to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Case page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CasePage"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          }
        }
      },
      "post": {
        "summary": "Create a Case, optionally with its first result",
        "parameters": [
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title"
                ],
                "additionalProperties": false,
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "note": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 5000
                  },
                  "track_id": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          }
        }
      }
    },
    "/cases/{case_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CaseId"
        }
      ],
      "get": {
        "summary": "Case detail with saved results",
        "responses": {
          "200": {
            "description": "Case detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "summary": "Update title and note",
        "parameters": [
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "version"
                ],
                "additionalProperties": false,
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "note": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 5000
                  },
                  "version": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated Case",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseSummary"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          }
        }
      }
    },
    "/cases/{case_id}/results": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CaseId"
        }
      ],
      "post": {
        "summary": "Save a search result into the Case (duplicates allowed)",
        "parameters": [
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "track_id"
                ],
                "additionalProperties": false,
                "properties": {
                  "track_id": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResult"
                }
              }
            }
          },
          "403": {
            "description": "forbidden or track_not_savable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/cases/{case_id}/results/{case_result_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CaseId"
        },
        {
          "$ref": "#/components/parameters/CaseResultId"
        }
      ],
      "delete": {
        "summary": "Remove a saved result; the track itself is kept",
        "parameters": [
          {
            "$ref": "#/components/parameters/Csrf"
          }
        ],
        "responses": {
          "204": {
            "description": "Removed"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/cases/{case_id}/results/{case_result_id}/crop": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CaseId"
        },
        {
          "$ref": "#/components/parameters/CaseResultId"
        }
      ],
      "get": {
        "summary": "Person crop of a saved result",
        "responses": {
          "200": {
            "description": "JPEG image",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "410": {
            "$ref": "#/components/responses/Gone"
          }
        }
      }
    },
    "/cases/{case_id}/results/{case_result_id}/frame": {
      "parameters": [
        {
          "$ref": "#/components/parameters/CaseId"
        },
        {
          "$ref": "#/components/parameters/CaseResultId"
        }
      ],
      "get": {
        "summary": "Full frame of a saved result",
        "responses": {
          "200": {
            "description": "JPEG image",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "410": {
            "$ref": "#/components/responses/Gone"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ps_session"
      }
    },
    "parameters": {
      "Csrf": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "CaseId": {
        "name": "case_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "CaseResultId": {
        "name": "case_result_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "schemas": {
      "CaseSummary": {
        "type": "object",
        "required": [
          "id",
          "title",
          "note",
          "owner",
          "result_count",
          "created_at",
          "updated_at",
          "version"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "owner": {
            "type": "object",
            "required": [
              "id",
              "display_name",
              "status"
            ],
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "display_name": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "LOCKED",
                  "INACTIVE",
                "DELETED"
                ]
              }
            }
          },
          "result_count": {
            "type": "integer",
            "minimum": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "CasePage": {
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CaseSummary"
            }
          },
          "next_cursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CaseResult": {
        "type": "object",
        "required": [
          "id",
          "case_id",
          "track_id",
          "camera_name",
          "area_name",
          "appeared_at",
          "saved_at",
          "bbox",
          "crop_url",
          "frame_url"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "case_id": {
            "type": "string",
            "format": "uuid"
          },
          "track_id": {
            "type": "string",
            "format": "uuid"
          },
          "camera_name": {
            "type": "string",
            "description": "Snapshot at save time"
          },
          "area_name": {
            "type": "string",
            "description": "Snapshot at save time"
          },
          "appeared_at": {
            "type": "string",
            "format": "date-time"
          },
          "saved_at": {
            "type": "string",
            "format": "date-time"
          },
          "bbox": {
            "type": "object",
            "nullable": true,
            "properties": {
              "x": {
                "type": "integer"
              },
              "y": {
                "type": "integer"
              },
              "width": {
                "type": "integer"
              },
              "height": {
                "type": "integer"
              },
              "frame_width": {
                "type": "integer"
              },
              "frame_height": {
                "type": "integer"
              }
            }
          },
          "crop_url": {
            "type": "string"
          },
          "frame_url": {
            "type": "string"
          }
        }
      },
      "CaseDetail": {
        "type": "object",
        "required": [
          "case",
          "results"
        ],
        "properties": {
          "case": {
            "$ref": "#/components/schemas/CaseSummary"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CaseResult"
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              },
              "request_id": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "responses": {
      "Forbidden": {
        "description": "forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "case_not_found or case_result_not_found (also when the Case belongs to another Operator)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "version_conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Validation": {
        "description": "validation_failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Gone": {
        "description": "image_unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}